Add typing to params
The idea here would be ensure more rigor when it comes to source params.
The proposal here would require devs to implement a params function that
conforms to the annotated types when they build their Source class.

The downside, I am not sure we can define the function for users,
however, at the very least, most can just implement with something that
returns an empty object